+++ ziti/cmd/ziti/cmd/fabric/create_router.go
@@ -29,11 +29,10 @@ import (
 
 type createRouterOptions struct {
 	api.Options
-	name           string
-	cost           uint16
-	tags           map[string]string
-	noTraversal    bool
-	allowTraversal bool
+	name        string
+	cost        uint16
+	tags        map[string]string
+	noTraversal bool
 }
 
 // newCreateRouterCmd creates the 'fabric create router' command for the given entity type
@@ -58,8 +57,7 @@ func newCreateRouterCmd(p common.OptionsProvider) *cobra.Command {
 	cmd.Flags().StringToStringVarP(&options.tags, "tags", "t", nil, "Add tags to router definition")
 	cmd.Flags().StringVar(&options.name, "name", "", "Specifies the router name. If not specified, the id in the controller cert will be used")
 	cmd.Flags().Uint16Var(&options.cost, "cost", 0, "Specifies the router cost. Default 0.")
-	cmd.Flags().BoolVar(&options.noTraversal, "disallow-traversal", false, "Disallow traversal for this edge router. Default to allowed. Mutually Exclusive to allow-traversal, but disallow has priority.")
-	cmd.Flags().BoolVar(&options.allowTraversal, "allow-traversal", false, "Allow traversal for this edge router. Default to allowed. Mutually Exclusive to disallow-traversal, but disallow has priority.")
+	cmd.Flags().BoolVar(&options.noTraversal, "no-traversal", false, "Disallow traversal for this edge router. Default to allowed(false).")
 
 	options.AddCommonFlags(cmd)
 
@@ -85,11 +83,7 @@ func (o *createRouterOptions) createRouter(_ *cobra.Command, args []string) erro
 	api.SetJSONValue(entityData, fingerprint, "fingerprint")
 	api.SetJSONValue(entityData, o.tags, "tags")
 	api.SetJSONValue(entityData, o.cost, "cost")
-	if o.noTraversal {
-		api.SetJSONValue(entityData, o.noTraversal, "noTraversal")
-	} else {
-		api.SetJSONValue(entityData, !o.allowTraversal, "noTraversal")
-	}
+	api.SetJSONValue(entityData, o.noTraversal, "noTraversal")
 	result, err := createEntityOfType("routers", entityData.String(), &o.Options)
 	if err != nil {
 		return err