+++ ziti/cmd/ziti/cmd/fabric/update_router.go
@@ -31,11 +31,10 @@ import (
 
 type updateRouterOptions struct {
 	api.Options
-	name           string
-	fingerprint    string
-	cost           uint16
-	noTraversal    bool
-	allowTraversal bool
+	name        string
+	fingerprint string
+	cost        uint16
+	noTraversal bool
 }
 
 func newUpdateRouterCmd(p common.OptionsProvider) *cobra.Command {
@@ -61,8 +60,7 @@ func newUpdateRouterCmd(p common.OptionsProvider) *cobra.Command {
 	cmd.Flags().StringVarP(&options.name, "name", "n", "", "Set the router name")
 	cmd.Flags().StringVar(&options.fingerprint, "fingerprint", "", "Sets the router fingerprint")
 	cmd.Flags().Uint16Var(&options.cost, "cost", 0, "Specifies the router cost. Default 0.")
-	cmd.Flags().BoolVar(&options.noTraversal, "disallow-traversal", false, "Disallow traversal for this edge router. Default to allowed. Mutually Exclusive to allow-traversal, but disallow has priority.")
-	cmd.Flags().BoolVar(&options.allowTraversal, "allow-traversal", false, "Allow traversal for this edge router. Default to allowed. Mutually Exclusive to disallow-traversal, but disallow has priority.")
+	cmd.Flags().BoolVar(&options.noTraversal, "no-traversal", false, "Disallow traversal for this edge router. Default to allowed(false).")
 
 	options.AddCommonFlags(cmd)
 
@@ -93,12 +91,7 @@ func runUpdateRouter(o *updateRouterOptions) error {
 		change = true
 	}
 
-	if o.Cmd.Flags().Changed("allow-traversal") {
-		api.SetJSONValue(entityData, !o.allowTraversal, "noTraversal")
-		change = true
-	}
-
-	if o.Cmd.Flags().Changed("disallow-traversal") {
+	if o.Cmd.Flags().Changed("no-traversal") {
 		api.SetJSONValue(entityData, o.noTraversal, "noTraversal")
 		change = true
 	}